+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 * Changed
   * Password field changed to password type.
   * Configuration variable for update URI moved to _ext_localconf.php_.
+  * Composer cache directory set.
 
 ## [1.1.1] - 2021-11-07
 * Fixed